Started porting AlarmViewControlle to Eureka forms (above/below sliders)
