+++ nightguard/Alarm.storyboard
@@ -19,7 +19,7 @@
                         <autoresizingMask key="autoresizingMask"/>
                     </navigationBar>
                     <connections>
-                        <segue destination="C5X-J9-KPA" kind="relationship" relationship="rootViewController" id="o2e-3g-ySK"/>
+                        <segue destination="1hI-q3-Ja3" kind="relationship" relationship="rootViewController" id="71r-dz-BfA"/>
                     </connections>
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="5Ux-Sh-UM6" userLabel="First Responder" sceneMemberID="firstResponder"/>
@@ -29,13 +29,13 @@
         <!--Alarm-->
         <scene sceneID="6Si-zS-Lu7">
             <objects>
-                <viewController wantsFullScreenLayout="YES" id="C5X-J9-KPA" customClass="AlarmViewController" customModule="nightguard" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="AlarmViewControllerOLD" wantsFullScreenLayout="YES" id="C5X-J9-KPA" customClass="AlarmViewControllerOLD" customModule="nightguard" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="2LD-TD-a5X">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="867"/>
                         <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                         <subviews>
                             <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="yDV-I8-xa6">
-                                <rect key="frame" x="0.0" y="64" width="375" height="803"/>
+                                <rect key="frame" x="0.0" y="20" width="375" height="847"/>
                                 <subviews>
                                     <view opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="hsU-pg-n3U">
                                         <rect key="frame" x="0.0" y="0.0" width="375" height="621"/>
@@ -353,7 +353,25 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="bgz-gv-tFu" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="980" y="-894.60269865067471"/>
+            <point key="canvasLocation" x="988" y="-54"/>
+        </scene>
+        <!--Alarms-->
+        <scene sceneID="ecR-Qg-ept">
+            <objects>
+                <viewController id="1hI-q3-Ja3" customClass="AlarmViewController" customModule="nightguard" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="hei-n6-4s8">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <viewLayoutGuide key="safeArea" id="2JQ-QA-vAB"/>
+                    </view>
+                    <tabBarItem key="tabBarItem" title="Alarms" image="Alarm" id="5WR-Lr-3ax"/>
+                    <navigationItem key="navigationItem" title="Alarms" id="7i7-rr-MK5"/>
+                    <simulatedTabBarMetrics key="simulatedBottomBarMetrics"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="TIJ-pQ-ukE" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="1022" y="-896"/>
         </scene>
     </scenes>
     <resources>
